Mule mostly working w/ exception to the IMU
All Talons are now registered in the the robot's drive subsystem and
working properly. Encoder are now programmer and will be wired later.
Encoders have been tested and do work. Further testing may be needed to
ensure they work properly. The IMU is reporting inaccurate data,
possibly due to being vertically mounted. Will have to test to see if
horizontal mounting will improve performance or if the IMU is dead (not
working right).