Accuracy Assured
IMU is now horizontally mounted on a piece of 2x4 on the vertical
electronics board. This has fix inaccuracies in reading rotational data
(Hailey believes the gyro's sensor may be accelerometer-based). Also,
the right encoder showed more signs of failure to signal. Turns out it
was a faulty wire that was replaced. Mule is now completely operational.