Documentation, Clean Up, and A small deletion
I simply added documentation to the DriveSubystem and commandd, as well
as cleaned up cluttered code in places. I also deleted the OI class (we
never used it anyways ¯\_(ツ)_/¯) and got rid of all its related code
elsewhere. We'll see if it breaks anything FIRST specific. If so, I
guess we add it back...