PID enhancements
Ravi came in and helped us solve our jerking issues when including D. This was solved by adding the same wrapping solution to derivative as we used on error. Also, PID is no long a feature of the GyroAutonomousTestCommand class and is its own class. Integral has been simplified back to simply adding trapeziod (hasn't been tested; will on Saturday).