PID code and turning
There is now code to travel a given distance using strictly PID. We tuned the values to have a fairly decent "PD" loop, but when "I" was introduced, we got wildly ranging "D" output values, causing the robot to violently jitter. I'm currently research to see why this is and how to fix it. If we can get a good PID only implementation for motion, hopefully we can go back to motionprofiling and see what went wrong (also Omari and Ben is looping at some of FIRST's give PID/Feed-foward tools).