Interpolation Work
Today we started working on cubic interpolation between two points. I
made a Python script that crunches the numbers for the coefficients, the
robot reads them from a file on a plugged in USB drive, and then it does
the math for the points it needs to travel to along the curve. Right now
none of our auto is working and I'm trying to figure out why....